Reorganize the service API
