+++ betty/service/provider.py
@@ -1,11 +1,13 @@
 """
-An API for providing application-wide services.
+Service providers.
 """
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-from collections.abc import Awaitable, Callable, Iterable, MutableSequence
+from _warnings import warn
+from abc import abstractmethod
+from collections.abc import Awaitable, Callable
+from functools import update_wrapper
 from inspect import getmembers, iscoroutinefunction
 from typing import (
     TYPE_CHECKING,
@@ -14,150 +16,36 @@ from typing import (
     Protocol,
     Self,
     TypeAlias,
-    TypedDict,
     TypeVar,
-    Unpack,
     cast,
     final,
     overload,
 )
-from warnings import warn
 
 from typing_extensions import override
 
 from betty.concurrent import AsynchronizedLock, Lock
 from betty.config import Configurable
-from betty.typing import Void, internal, public, unpickleable
+from betty.service import ServiceError
+from betty.service.bootstrap import Bootstrapped, Shutdownable, ShutdownStack
+from betty.typing import Void, internal, public
 
 if TYPE_CHECKING:
+    from collections.abc import Iterable
     from types import TracebackType
 
 
-@internal
-class ServiceError(RuntimeError):
-    """
-    A service API error.
-    """
+_ServiceT = TypeVar("_ServiceT")
+_ServiceGetT = TypeVar("_ServiceGetT")
 
 
 @internal
-class BootstrappedError(ServiceError):
-    """
-    Something was unexpectedly bootstrapped already.
-    """
-
-
-@internal
-class NotBootstrappedError(ServiceError):
-    """
-    Something was unexpectedly not yet bootstrapped.
-    """
-
-
-@internal
-class ServiceInitializedError(ServiceError):
-    """
-    A service was unexpectedly initialized already.
-    """
-
-
-@internal
-class Bootstrapped:
-    """
-    A component that can be in a bootstrapped state.
-    """
-
-    def __init__(self, *args: Any, **kwargs: Any):
-        self._bootstrapped = False
-        super().__init__(*args, **kwargs)
-
-    @final
-    @property
-    def bootstrapped(self) -> bool:
-        """
-        Whether the component has been bootstrapped.
-        """
-        return self._bootstrapped
-
-    @final
-    def assert_bootstrapped(self) -> None:
-        """
-        Assert that the component has been bootstrapped.
-        """
-        if not self.bootstrapped:
-            raise NotBootstrappedError(f"{self} was not bootstrapped yet.")
-
-    @final
-    def assert_not_bootstrapped(self) -> None:
-        """
-        Assert that the component was not bootstrapped.
-        """
-        if self.bootstrapped:
-            raise BootstrappedError(f"{self} was bootstrapped already.")
-
-
-class Shutdownable(ABC):
-    """
-    A component that can be shut down.
-    """
-
-    @abstractmethod
-    async def shutdown(self, *, wait: bool = True) -> None:
-        """
-        Shut the component down.
-        """
-
-
-class ShutdownCallbackKwargs(TypedDict):
-    """
-    The keyword arguments to a shutdown callback.
-    """
-
-    #: ``True`` to wait for the component to shut down gracefully, or ``False`` to attempt an immediate forced shutdown.
-    wait: bool
-
-
-ShutdownCallback: TypeAlias = Callable[
-    [Unpack[ShutdownCallbackKwargs]], Awaitable[None]
-]
-
-
-@internal
-@final
-class ShutdownStack(Bootstrapped, Shutdownable):
-    """
-    A stack that invokes callbacks in reverse order upon shutting down.
-    """
-
-    def __init__(self):
-        super().__init__()
-        self._bootstrapped = True
-        self._callbacks: MutableSequence[ShutdownCallback] = []
-
-    @override
-    async def shutdown(self, *, wait: bool = True) -> None:
-        self.assert_bootstrapped()
-        self._bootstrapped = False
-        for callback in reversed(self._callbacks):
-            await callback(wait=wait)
-
-    def append(self, callback: ShutdownCallback | Shutdownable) -> None:
-        """
-        Append a callback or another component to the stack.
-        """
-        self._callbacks.append(
-            callback.shutdown if isinstance(callback, Shutdownable) else callback
-        )
-
-
-@internal
-@unpickleable
 class ServiceProvider(Bootstrapped, Shutdownable):
     """
     A service provider.
 
     Service providers make up a running Betty 'application'. They can provide services through
-    :py:func:`betty.service.service`, and manage their resources by being bootstrapped and shut down.
+    :py:func:`betty.service.provider.service`, and manage their resources by being bootstrapped and shut down.
     """
 
     def __init__(self, *args: Any, **kwargs: Any):
@@ -217,10 +105,82 @@ class ServiceProvider(Bootstrapped, Shutdownable):
 
 
 _ServiceProviderT = TypeVar("_ServiceProviderT", bound=ServiceProvider)
-_ServiceT = TypeVar("_ServiceT")
-_ServiceGetT = TypeVar("_ServiceGetT")
+
 
 ServiceFactory: TypeAlias = Callable[[_ServiceProviderT], _ServiceT]
+
+
+class _ServiceDecorator(Protocol):
+    @overload
+    def __call__(
+        self, factory: Callable[[_ServiceProviderT], _ServiceT]
+    ) -> _SynchronousServiceManager[_ServiceProviderT, _ServiceT]:
+        pass
+
+    @overload
+    def __call__(
+        self, factory: Callable[[_ServiceProviderT], Awaitable[_ServiceT]]
+    ) -> _AsynchronousServiceManager[_ServiceProviderT, _ServiceT]:
+        pass
+
+
+@overload
+def service(  # type: ignore[overload-overlap]
+    factory: Callable[[_ServiceProviderT], Awaitable[_ServiceT]],
+) -> _AsynchronousServiceManager[_ServiceProviderT, _ServiceT]:
+    pass
+
+
+@overload
+def service(
+    factory: Callable[[_ServiceProviderT], _ServiceT],
+) -> _SynchronousServiceManager[_ServiceProviderT, _ServiceT]:
+    pass
+
+
+@overload
+def service(factory: None = None) -> _ServiceDecorator:
+    pass
+
+
+def service(
+    factory: Callable[[_ServiceProviderT], _ServiceGetT] | None = None,
+) -> ServiceManager[_ServiceProviderT, _ServiceGetT, Any] | _ServiceDecorator:
+    """
+    Decorate a service factory method.
+
+    The factory method is replaced with a :py:class:`service manager <betty.service.provider.ServiceManager>` which
+    handles lazy service instantiation, caching, and multiprocessing support.
+
+    The decorated factory method should return a new service instance.
+    """
+
+    def _service(
+        factory: Callable[[_ServiceProviderT], _ServiceGetT],
+    ) -> ServiceManager[_ServiceProviderT, _ServiceGetT, Any]:
+        if iscoroutinefunction(factory):
+            return _AsynchronousServiceManager(factory)  # type: ignore[return-value]
+        return _SynchronousServiceManager(factory)
+
+    if factory is None:
+        return _service  # type: ignore[return-value]
+    return _service(factory)
+
+
+@internal
+class StaticService(Generic[_ServiceProviderT, _ServiceT]):
+    """
+    A service factory that returns a static, predefined service.
+    """
+
+    def __init__(self, service: _ServiceT):
+        self._service = service
+
+    def __call__(self, services: _ServiceProviderT) -> _ServiceT:
+        """
+        Return the service.
+        """
+        return self._service
 
 
 @internal
@@ -230,6 +190,10 @@ class ServiceManager(Generic[_ServiceProviderT, _ServiceGetT, _ServiceT]):
     """
 
     def __init__(self, factory: ServiceFactory[_ServiceProviderT, _ServiceGetT]):
+        update_wrapper(  # type: ignore[type-var]
+            self,
+            factory,
+        )
         self._factory = factory
         self._service_name: str = factory.__name__  # type: ignore[attr-defined]
         self._service_attr_name = f"_{self._service_name}"
@@ -365,74 +329,8 @@ class _SynchronousServiceManager(
         return new_service
 
 
-class _ServiceDecorator(Protocol):
-    @overload
-    def __call__(
-        self, factory: Callable[[_ServiceProviderT], _ServiceT]
-    ) -> _SynchronousServiceManager[_ServiceProviderT, _ServiceT]:
-        pass
-
-    @overload
-    def __call__(
-        self, factory: Callable[[_ServiceProviderT], Awaitable[_ServiceT]]
-    ) -> _AsynchronousServiceManager[_ServiceProviderT, _ServiceT]:
-        pass
-
-
-@overload
-def service(  # type: ignore[overload-overlap]
-    factory: Callable[[_ServiceProviderT], Awaitable[_ServiceT]],
-) -> _AsynchronousServiceManager[_ServiceProviderT, _ServiceT]:
-    pass
-
-
-@overload
-def service(
-    factory: Callable[[_ServiceProviderT], _ServiceT],
-) -> _SynchronousServiceManager[_ServiceProviderT, _ServiceT]:
-    pass
-
-
-@overload
-def service(factory: None = None) -> _ServiceDecorator:
-    pass
-
-
-def service(
-    factory: Callable[[_ServiceProviderT], _ServiceGetT] | None = None,
-) -> ServiceManager[_ServiceProviderT, _ServiceGetT, Any] | _ServiceDecorator:
-    """
-    Decorate a service factory method.
-
-    The factory method is replaced with a :py:class:`service manager <betty.service.ServiceManager>` which handles lazy
-    service instantiation, caching, and multiprocessing support.
-
-    The decorated factory method should return a new service instance.
-    """
-
-    def _service(
-        factory: Callable[[_ServiceProviderT], _ServiceGetT],
-    ) -> ServiceManager[_ServiceProviderT, _ServiceGetT, Any]:
-        if iscoroutinefunction(factory):
-            return _AsynchronousServiceManager(factory)  # type: ignore[return-value]
-        return _SynchronousServiceManager(factory)
-
-    if factory is None:
-        return _service  # type: ignore[return-value]
-    return _service(factory)
-
-
 @internal
-class StaticService(Generic[_ServiceProviderT, _ServiceT]):
+class ServiceInitializedError(ServiceError):
     """
-    A service factory that returns a static, predefined service.
+    A service was unexpectedly initialized already.
     """
-
-    def __init__(self, service: _ServiceT):
-        self._service = service
-
-    def __call__(self, service_provider: _ServiceProviderT) -> _ServiceT:
-        """
-        Return the service.
-        """
-        return self._service
